requirements.txt
asgiref==3.8.1
astroid==3.2.2
dill==0.3.8
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
Faker==25.2.0
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
platformdirs==4.2.2
pylint==3.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.11.0
validate-docbr==1.10.0
wrapt==1.16.0
